+++ Branch-SDK/Branch.m
@@ -45,6 +45,7 @@
 #import "BNCAppGroupsData.h"
 #import "BNCPartnerParameters.h"
 #import "BranchEvent.h"
+#import "BNCPasteboard.h"
 
 #if !TARGET_OS_TV
 #import "BNCUserAgentCollector.h"
@@ -224,6 +225,10 @@ typedef NS_ENUM(NSInteger, BNCInitStatus) {
     [self loadUserAgent];
     
     BranchJsonConfig *config = BranchJsonConfig.instance;
+    
+    if (config.checkPasteboardOnInstall) {
+        [self checkPasteboardOnInstall];
+    }
     
     if (config.delayInitToCheckForSearchAds) {
         [self delayInitToCheckForSearchAds];
@@ -800,26 +805,7 @@ static BOOL bnc_enableFingerprintIDInCrashlyticsReports = YES;
 
     [self initUserSessionAndCallCallback:YES sceneIdentifier:sceneIdentifier];
 
-    id branchUniversalLinkDomains = [self.preferenceHelper getBranchUniversalLinkDomains];
-    if ([branchUniversalLinkDomains isKindOfClass:[NSString class]] && [urlString bnc_containsString:branchUniversalLinkDomains]) {
-        return YES;
-    } else if ([branchUniversalLinkDomains isKindOfClass:[NSArray class]]) {
-        for (id oneDomain in branchUniversalLinkDomains) {
-            if ([oneDomain isKindOfClass:[NSString class]] && [urlString bnc_containsString:oneDomain]) {
-                return YES;
-            }
-        }
-    }
-
-    NSString *userActivityURL = urlString;
-    NSArray *branchDomains = [NSArray arrayWithObjects:@"bnc.lt", @"app.link", @"test-app.link", nil];
-    for (NSString* domain in branchDomains) {
-        if ([userActivityURL bnc_containsString:domain]) {
-            return YES;
-        }
-    }
-
-    return NO;
+    return [Branch isBranchLink:urlString];
 }
 
 - (BOOL)continueUserActivity:(NSUserActivity *)userActivity {
@@ -829,6 +815,12 @@ static BOOL bnc_enableFingerprintIDInCrashlyticsReports = YES;
 - (BOOL)continueUserActivity:(NSUserActivity *)userActivity sceneIdentifier:(NSString *)sceneIdentifier {
     BNCLogDebugSDK(@"continueUserActivity:");
 
+    if (@available(iOS 11.0, *)) {
+        if (userActivity.referrerURL) {
+            self.preferenceHelper.initialReferrer = userActivity.referrerURL.absoluteString;
+        }
+    }
+    
     // Check to see if a browser activity needs to be handled
     if ([userActivity.activityType isEqualToString:NSUserActivityTypeBrowsingWeb]) {
         return [self handleDeepLink:userActivity.webpageURL sceneIdentifier:sceneIdentifier];
@@ -841,9 +833,9 @@ static BOOL bnc_enableFingerprintIDInCrashlyticsReports = YES;
     spotlightIdentifier = [self.contentDiscoveryManager spotlightIdentifierFromActivity:userActivity];
     NSURL *webURL = userActivity.webpageURL;
 
-    if ([self isBranchLink:userActivity.userInfo[CSSearchableItemActivityIdentifier]]) {
+    if ([Branch isBranchLink:userActivity.userInfo[CSSearchableItemActivityIdentifier]]) {
         return [self handleDeepLink:[NSURL URLWithString:userActivity.userInfo[CSSearchableItemActivityIdentifier]] sceneIdentifier:sceneIdentifier];
-    } else if (webURL != nil && [self isBranchLink:[webURL absoluteString]]) {
+    } else if (webURL != nil && [Branch isBranchLink:[webURL absoluteString]]) {
         return [self handleDeepLink:webURL sceneIdentifier:sceneIdentifier];
     } else if (spotlightIdentifier) {
         self.preferenceHelper.spotlightIdentifier = spotlightIdentifier;
@@ -860,13 +852,14 @@ static BOOL bnc_enableFingerprintIDInCrashlyticsReports = YES;
     return spotlightIdentifier != nil;
 }
 
-- (BOOL)isBranchLink:(NSString*)urlString {
-    id branchUniversalLinkDomains = [self.preferenceHelper getBranchUniversalLinkDomains];
-    if ([branchUniversalLinkDomains isKindOfClass:[NSString class]] &&
-        [urlString containsString:branchUniversalLinkDomains]) {
+// checks if URL string looks like a branch link
++ (BOOL)isBranchLink:(NSString *)urlString {
+    id branchUniversalLinkDomains = [[[NSBundle mainBundle] infoDictionary] objectForKey:@"branch_universal_link_domains"];
+    
+    // check url list in bundle
+    if ([branchUniversalLinkDomains isKindOfClass:[NSString class]] && [urlString containsString:branchUniversalLinkDomains]) {
         return YES;
-    }
-    else if ([branchUniversalLinkDomains isKindOfClass:[NSArray class]]) {
+    } else if ([branchUniversalLinkDomains isKindOfClass:[NSArray class]]) {
         for (id oneDomain in branchUniversalLinkDomains) {
             if ([oneDomain isKindOfClass:[NSString class]] && [urlString containsString:oneDomain]) {
                 return YES;
@@ -874,11 +867,13 @@ static BOOL bnc_enableFingerprintIDInCrashlyticsReports = YES;
         }
     }
 
+    // check default urls
     NSString *userActivityURL = urlString;
     NSArray *branchDomains = [NSArray arrayWithObjects:@"bnc.lt", @"app.link", @"test-app.link", nil];
     for (NSString* domain in branchDomains) {
-        if ([userActivityURL containsString:domain])
+        if ([userActivityURL containsString:domain]) {
             return YES;
+        }
     }
     return NO;
 }
@@ -948,6 +943,19 @@ static BOOL bnc_enableFingerprintIDInCrashlyticsReports = YES;
     #if !TARGET_OS_TV
     [BNCAppleSearchAds sharedInstance].ignoreAppleTestData = YES;
     #endif
+}
+
+- (void)checkPasteboardOnInstall {
+    [BNCPasteboard sharedInstance].checkOnInstall = YES;
+}
+
+- (BOOL)willShowPasteboardToast {
+    if (!self.preferenceHelper.identityID &&
+        [BNCPasteboard sharedInstance].checkOnInstall &&
+        [BNCPasteboard sharedInstance].isUrlOnPasteboard) {
+        return YES;
+    }
+    return NO;
 }
 
 - (void)checkAppleSearchAdsAttribution {