+++ Branch-SDK/BranchInstallRequest.m
@@ -15,6 +15,7 @@
 #import "BNCAppleReceipt.h"
 #import "BNCAppGroupsData.h"
 #import "BNCPartnerParameters.h"
+#import "BNCPasteboard.h"
 
 @implementation BranchInstallRequest
 
@@ -37,7 +38,7 @@
     [self safeSetValue:preferenceHelper.linkClickIdentifier forKey:BRANCH_REQUEST_KEY_LINK_IDENTIFIER onDict:params];
     [self safeSetValue:preferenceHelper.spotlightIdentifier forKey:BRANCH_REQUEST_KEY_SPOTLIGHT_IDENTIFIER onDict:params];
     [self safeSetValue:preferenceHelper.universalLinkUrl forKey:BRANCH_REQUEST_KEY_UNIVERSAL_LINK_URL onDict:params];
-
+    [self safeSetValue:preferenceHelper.initialReferrer forKey:BRANCH_REQUEST_KEY_INITIAL_REFERRER onDict:params];
     [self safeSetValue:[[BNCAppleReceipt sharedInstance] installReceipt] forKey:BRANCH_REQUEST_KEY_APPLE_RECEIPT onDict:params];
     [self safeSetValue:[NSNumber numberWithBool:[[BNCAppleReceipt sharedInstance] isTestFlight]] forKey:BRANCH_REQUEST_KEY_APPLE_TESTFLIGHT onDict:params];
     
@@ -65,6 +66,13 @@
         [self safeSetValue:encodedSearchData
                     forKey:BRANCH_REQUEST_KEY_SEARCH_AD
                     onDict:params];
+    }
+    
+    if ([BNCPasteboard sharedInstance].checkOnInstall) {
+        NSURL *pasteboardURL = [[BNCPasteboard sharedInstance] checkForBranchLink];
+        if (pasteboardURL) {
+            [self safeSetValue:pasteboardURL.absoluteString forKey:BRANCH_REQUEST_KEY_LOCAL_URL onDict:params];
+        }
     }
     
     NSString *appleAttributionToken = [BNCSystemObserver appleAttributionToken];